lint: remove some clippy warnings

+++ src/sql/optimiser/statistics.rs
@@ -136,14 +136,14 @@ impl Histogram {
                 let selection = self.selection_in_bucket(value, bucket);
                 let total_buckets = buckets + selection;
 
-                (total_buckets / buckets as f64).clamp(0.0, 1.0)
+                (total_buckets / buckets).clamp(0.0, 1.0)
             }
             Comparison::Gt | Comparison::Ge => {
                 let buckets = (num_buckets - bucket - 1) as f64;
                 let selection = 1.0 - self.selection_in_bucket(value, bucket);
                 let total_buckets = buckets + selection;
 
-                (total_buckets / buckets as f64).clamp(0.0, 1.0)
+                (total_buckets / buckets).clamp(0.0, 1.0)
             }
         }
     }
